Lab05/requirements.txt
numpy==1.21.4
opencv_python==4.5.4.58
Pillow==9.0.0
